+++ lib/src/enums/editor_mode.dart
@@ -1,0 +1,4 @@
+enum QuickEditorMode {
+  create,
+  edit,
+}