+++ lib/src/util/locale.dart
@@ -1,0 +1,5 @@
+import 'dart:ui';
+
+extension QuickStartLocaleIterableExtension on Iterable<Locale> {
+  Iterable<String> get languageCodes => map((locale) => locale.languageCode);
+}