+++ lib/src/util/map.dart
@@ -1,8 +1,0 @@
-import '../quickstart.dart';
-
-extension QuickMapStringStringExtension<T> on Map<String, T> {
-  T? quickTr() {
-    final lang = QuickStart.delegate.settingsNotifier.lang;
-    return this[lang];
-  }
-}