+++ lib/src/widgets/layout/scaffold.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:multi_animated_builder/multi_animated_builder.dart';
 
 import '../../delegate.dart';
 import '../../quickstart.dart';
@@ -14,8 +15,11 @@ import 'horizontal_black_line.dart';
 class QuickScaffold extends StatelessWidget {
   final Widget body;
   final BodySize bodySize;
+  final WidgetBuilder? builder;
   final QuickDelegate? delegate;
   final double? maxWidth;
+  final List<Listenable> notifiers;
+  final VoidCallback? onSavePressed;
   final bool padding;
 
   const QuickScaffold({
@@ -23,8 +27,20 @@ class QuickScaffold extends StatelessWidget {
     required this.bodySize,
     this.delegate,
     this.maxWidth,
+    this.onSavePressed,
     this.padding = true,
-  });
+  })  : builder = null,
+        notifiers = const [];
+
+  const QuickScaffold.builder({
+    required this.builder,
+    required this.bodySize,
+    this.delegate,
+    this.maxWidth,
+    this.notifiers = const [],
+    this.onSavePressed,
+    this.padding = true,
+  }) : body = const SizedBox.shrink();
 
   const QuickScaffold.loading({
     this.delegate,
@@ -33,7 +49,10 @@ class QuickScaffold extends StatelessWidget {
   })  : bodySize = BodySize.infinite,
         body = const Center(
           child: SmallCircularProgressIndicator(),
-        );
+        ),
+        builder = null,
+        notifiers = const [],
+        onSavePressed = null;
 
   @override
   Widget build(BuildContext context) {
@@ -41,6 +60,13 @@ class QuickScaffold extends StatelessWidget {
     final delegate = this.delegate ?? QuickStart.delegate;
 
     Widget child = body;
+
+    if (notifiers.isNotEmpty) {
+      child = MultiAnimatedBuilder(
+        animations: notifiers,
+        builder: builder!,
+      );
+    }
 
     if (maxWidth != null) {
       child = ConstrainedWidthWidget(
@@ -81,9 +107,15 @@ class QuickScaffold extends StatelessWidget {
                 child: Column(
                   children: [
                     if (isMobile)
-                      _NarrowHeader(delegate: delegate)
+                      _NarrowHeader(
+                        delegate: delegate,
+                        onSavePressed: onSavePressed,
+                      )
                     else
-                      _WideHeader(delegate: delegate),
+                      _WideHeader(
+                        delegate: delegate,
+                        onSavePressed: onSavePressed,
+                      ),
                     const HorizontalBlackLine(),
                     Expanded(
                       child: child,
@@ -102,9 +134,11 @@ class QuickScaffold extends StatelessWidget {
 
 class _NarrowHeader extends StatelessWidget {
   final QuickDelegate delegate;
+  final VoidCallback? onSavePressed;
 
   const _NarrowHeader({
     required this.delegate,
+    required this.onSavePressed,
   });
 
   @override
@@ -127,6 +161,8 @@ class _NarrowHeader extends StatelessWidget {
             Expanded(
               child: delegate.buildHeaderFiller(context),
             ),
+            if (onSavePressed != null)
+              _SaveButton(ext: ext, onPressed: onSavePressed!),
           ].intersperse(delegate.spacing).toList(growable: true),
         ),
       ),
@@ -136,9 +172,11 @@ class _NarrowHeader extends StatelessWidget {
 
 class _WideHeader extends StatelessWidget {
   final QuickDelegate delegate;
+  final VoidCallback? onSavePressed;
 
   const _WideHeader({
     required this.delegate,
+    required this.onSavePressed,
   });
 
   @override
@@ -159,15 +197,33 @@ class _WideHeader extends StatelessWidget {
                 showText: true,
               ),
             ),
-            const SizedBox(width: 20),
             Expanded(
               child: delegate.buildHeaderFiller(context),
             ),
-            const SizedBox(width: 20),
             delegate.buildMenu(context),
-          ],
+            if (onSavePressed != null)
+              _SaveButton(ext: ext, onPressed: onSavePressed!),
+          ].intersperse(const SizedBox(width: 20)).toList(growable: false),
         ),
       ),
+    );
+  }
+}
+
+class _SaveButton extends StatelessWidget {
+  final QuickThemeExtension ext;
+  final VoidCallback onPressed;
+
+  const _SaveButton({
+    required this.ext,
+    required this.onPressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      onPressed: onPressed,
+      icon: Icon(Icons.check, color: ext.onHeaderColor),
     );
   }
 }